# packages/contracts/openapi/public-api.openapi.yaml
openapi: 3.1.0
info:
  title: Hathor Public API
  version: 1.0.0-draft
  description: |
    Contract for the August 30 vertical slice. All authenticated identity derives from
    a verified bearer token. Monetary values are decimal strings in EGP.
servers:
  - url: https://api.hathor.example/api/v1
    description: Public staging placeholder
  - url: http://localhost:5000/api/v1
    description: Local gateway
...
paths:
...
  /store/games:
    get:
      tags: [Store]
      operationId: listGames
      security: []
      parameters:
        - { name: query, in: query, schema: { type: string, maxLength: 100 } }
        - { name: tag, in: query, schema: { type: string, maxLength: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 50, default: 20 },
          }
      responses:
        '200':
          {
            description: Published games,
            content: { application/json: { schema: { $ref: '#/components/schemas/GamePage' } } },
          }
        '400': { $ref: '#/components/responses/ValidationFailed' }
  /store/games/{slug}:
    get:
      tags: [Store]
      operationId: getGameBySlug
      security: []
      parameters:
        - { $ref: '#/components/parameters/Slug' }
      responses:
        '200':
          {
            description: Published game,
            content: { application/json: { schema: { $ref: '#/components/schemas/Game' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Slug:
      {
        name: slug,
        in: path,
        required: true,
        schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' },
      }
  responses:
...
    NotFound:
      {
        description: Resource not found,
        content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } },
      }
...
    ValidationFailed:
      {
        description: Invalid request,
        content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } },
      }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code: { type: string }
            message: { type: string }
            correlationId: { $ref: '#/components/schemas/Uuid' }
...
    Game:
      type: object
      required: [id, slug, title, priceEgp, currency, status, theme]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        title: { type: string }
        shortDescription: { type: string }
        priceEgp: { type: string, pattern: '^\d+\.\d{2}$' }
        currency: { type: string, const: EGP }
        status: { type: string, enum: [draft, pending_review, published, rejected, suspended] }
        theme: { $ref: '#/components/schemas/ThemeDocument' }
        genreId: { type: ['integer', 'null'] }
        genre:
          anyOf:
            - $ref: '#/components/schemas/Genre'
            - type: 'null'
    GamePage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Game' } }
        nextCursor: { type: [string, 'null'] }
    ThemeDocument:
      type: object
      additionalProperties: false
      required: [schemaVersion, palette, typography, layout, contentOrder]
      properties:
        schemaVersion: { type: integer, const: 1 }
        palette:
          type: object
          additionalProperties: false
          required: [background, surface, text, accent]
          properties:
            background: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
            surface: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
            text: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
            accent: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        typography:
          type: object
          additionalProperties: false
          required: [headingFont, bodyFont, headingScale]
          properties:
            headingFont: { type: string, enum: [inter, cairo] }
            bodyFont: { type: string, enum: [inter, cairo] }
            headingScale: { type: string, enum: [normal, large] }
        layout:
          type: object
          additionalProperties: false
          required: [template, heroAlignment, cardStyle]
          properties:
            template: { type: string, enum: [standard, cinematic, compact] }
            heroAlignment: { type: string, enum: [left, center] }
            cardStyle: { type: string, enum: [flat, elevated] }
            showTrailer: { type: boolean, default: true }
        contentOrder:
          type: array
          uniqueItems: true
          maxItems: 4
          items: { type: string, enum: [hero, description, screenshots, systemRequirements] }
...
    Genre:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 50 }
        slug: { type: string, maxLength: 50 }
...